Re-group api's based on alpha.4 event api update
