+++ core/payment/examples/payment_api.rs
@@ -21,6 +21,7 @@ use ya_persistence::executor::DbExecutor;
 use ya_service_api_web::middleware::auth::dummy::DummyAuth;
 use ya_service_api_web::middleware::Identity;
 use ya_service_api_web::rest_api_addr;
+use ya_service_api_web::scope::ExtendableScope;
 use ya_service_bus::typed as bus;
 use ya_zksync_driver as zksync;
 
@@ -293,14 +294,16 @@ async fn main() -> anyhow::Result<()> {
             role: "".to_string(),
         };
 
-        let provider_scope =
-            ya_payment::api::provider_scope().wrap(DummyAuth::new(provider_identity));
-        let requestor_scope =
-            ya_payment::api::requestor_scope().wrap(DummyAuth::new(requestor_identity));
+        let provider_api_scope = Scope::new("/provider")
+            .extend(ya_payment::api::api_scope)
+            .wrap(DummyAuth::new(provider_identity));
+        let requestor_api_scope = Scope::new("/requestor")
+            .extend(ya_payment::api::api_scope)
+            .wrap(DummyAuth::new(requestor_identity));
         let payment_service = Scope::new(PAYMENT_API_PATH)
             .data(db.clone())
-            .service(provider_scope)
-            .service(requestor_scope);
+            .service(provider_api_scope)
+            .service(requestor_api_scope);
         App::new()
             .wrap(middleware::Logger::default())
             .service(payment_service)
